Finally fix build
I had to get rid of my psake build script and add a build
script using NMake to get this to work correctly. In addition,
the trailing backslashes in the values of $(IntDir) and $(OutDir)
were still causing problems, so I had to cheat by adding a dot after
the end of the value -- thus turning ...\Debug into ...\Debug\.
which are functionally equivalent.